1. fixed memory leak 2. fixed the condition is always false 3. fixed no return statement in function 4. fixed converting double to float 5. updated .gitignore for CLion

+++ src/planning/scenario_planning/lane_driving/behavior_planning/lane_change_planner/src/lane_changer.cpp
@@ -253,9 +253,9 @@ std_msgs::ColorRGBA toRainbow(double ratio)
       break;  // magenta
   }
   std_msgs::ColorRGBA color;
-  color.r = static_cast<double>(red) / 255;
-  color.g = static_cast<double>(grn) / 255;
-  color.b = static_cast<double>(blu) / 255;
+  color.r = static_cast<float>(red) / 255.f;
+  color.g = static_cast<float>(grn) / 255.f;
+  color.b = static_cast<float>(blu) / 255.f;
   color.a = 0.3;
 
   return color;
